refactor: break out TokenDetailsWidget to reduce complexity of TokenDetails
