+++ src/pages/TokenDetails/index.tsx
@@ -16,20 +16,22 @@ import MobileBalanceSummaryFooter from 'components/Tokens/TokenDetails/MobileBal
 import StatsSection from 'components/Tokens/TokenDetails/StatsSection'
 import TokenSafetyMessage from 'components/TokenSafety/TokenSafetyMessage'
 import TokenSafetyModal from 'components/TokenSafety/TokenSafetyModal'
-import Widget, { Field, Tokens, WIDGET_WIDTH } from 'components/Widget'
+import { WIDGET_WIDTH } from 'components/Widget'
 import { isCelo, nativeOnChain } from 'constants/tokens'
 import { checkWarning } from 'constants/tokenSafety'
 import { Chain } from 'graphql/data/__generated__/TokenQuery.graphql'
 import { useTokenQuery } from 'graphql/data/Token'
-import { CHAIN_NAME_TO_CHAIN_ID, chainIdToBackendName, validateUrlChainParam } from 'graphql/data/util'
+import { CHAIN_NAME_TO_CHAIN_ID, validateUrlChainParam } from 'graphql/data/util'
 import { useIsUserAddedTokenOnChain } from 'hooks/Tokens'
 import { useOnGlobalChainSwitch } from 'hooks/useGlobalChainSwitch'
 import { useAtomValue } from 'jotai/utils'
 import useCurrencyBalance, { useTokenBalance } from 'lib/hooks/useCurrencyBalance'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useMemo, useState } from 'react'
 import { ArrowLeft } from 'react-feather'
 import { useNavigate, useParams } from 'react-router-dom'
 import styled from 'styled-components/macro'
+
+import TokenDetailsWidget from './TokenDetailsWidget'
 
 export const TokenDetailsLayout = styled.div`
   display: flex;
@@ -147,7 +149,6 @@ export default function TokenDetails() {
     [continueSwap, setContinueSwap]
   )
 
-  const [tokens, setTokens] = useState<Tokens>({})
   const widgetToken = useMemo(() => {
     if (pageToken) {
       return pageToken
@@ -156,23 +157,6 @@ export default function TokenDetails() {
     }
     return undefined
   }, [nativeCurrency, pageChainId, pageToken])
-  useEffect(() => {
-    if (tokens[Field.INPUT] === widgetToken || tokens[Field.OUTPUT] === widgetToken) return
-    setTokens({ [Field.OUTPUT]: widgetToken })
-  }, [tokens, widgetToken])
-  const onTokensChange = useCallback(
-    (tokens: Tokens) => {
-      setTokens(tokens)
-      if (tokens[Field.INPUT] === widgetToken || tokens[Field.OUTPUT] === widgetToken) return
-      const token = tokens[Field.OUTPUT] || tokens[Field.INPUT]
-      if (!token) return
-
-      const network = chainIdToBackendName(token.chainId).toLowerCase()
-      const address = token.isNative ? 'NATIVE' : token.address
-      navigate(`/tokens/${network}/${address}`)
-    },
-    [navigate, widgetToken]
-  )
 
   return (
     <TokenDetailsLayout>
@@ -202,7 +186,7 @@ export default function TokenDetails() {
             <AddressSection address={tokenQueryData.address ?? ''} />
           </LeftPanel>
           <RightPanel>
-            <Widget tokens={tokens} onTokensChange={onTokensChange} onReviewSwapClick={onReviewSwap} />
+            <TokenDetailsWidget token={widgetToken} onReviewSwap={onReviewSwap} />
             {tokenWarning && <TokenSafetyMessage tokenAddress={tokenQueryData.address ?? ''} warning={tokenWarning} />}
             <BalanceSummary
               tokenAmount={tokenBalance}
